requirements.txt
flask
flask-restful
pillow
gunicorn
